fix my wifi-connect script.
When it is detected that the wpa_supplicant status is COMPLETED,
wifi-connect exits immediately without notifying i3blocks to update.
This may cause i3blocks to fail to reflect the actual state of
wpa_supplicant in a timely manner. Therefore, after the while loop ends,
execute `pkill i3blocks` to notify i3blocks to update.