address-query: improve POSIX-compliance
POSIX `head` does not support the use of a plus sign ('+') preceding the
number in the argument of the `-n` option. To address this limitation,
the following changes are applied:

.config/neomutt/scripts/address-query: Use `tail` instead of `head`.
documents/addressbook.gpg: Move the Vim modeline from the last line to
the first line to make the command `tail` work.